+++ lib/dvb/tstools.cpp
@@ -124,7 +124,7 @@ int eDVBTSTools::getPTS(off_t &offset, pts_t &pts, int fixed)
 						// obsolete data that happens to have a '1' there
 						continue;
 					}
-					eDebug("[eDVBTSTools] getPTS got it from sc file offset=%ld pts=%lld", (intmax_t)local_offset, pts);
+					eDebug("[eDVBTSTools] getPTS got it from sc file offset=%lld pts=%lld", (long long)local_offset, pts);
 					if (fixed && fixupPTS(local_offset, pts))
 					{
 						eDebug("[eDVBTSTools]    But failed to fixup!");
@@ -170,7 +170,7 @@ int eDVBTSTools::getPTS(off_t &offset, pts_t &pts, int fixed)
 			const unsigned char* match = (const unsigned char*)memchr(packet+1, 0x47, 188-1);
 			if (match != NULL)
 			{
-				eDebug("[eDVBTSTools] getPTS resync %ld", match - packet);
+				eDebug("[eDVBTSTools] getPTS resync %td", match - packet);
 				offset += (match - packet) - 188;
 			}
 			else
@@ -208,8 +208,8 @@ int eDVBTSTools::getPTS(off_t &offset, pts_t &pts, int fixed)
 					pts |= ((unsigned long long)(packet[ 9]&0xFF)) << 1;
 					pts |= ((unsigned long long)(packet[10]&0x80)) >> 7;
 					offset -= 188;
-					eDebug("[eDVBTSTools] getPTS PCR %16llx found at %ld pid %02x (%02x %02x %02x %02x %02x %02x %02x %02x %02x %02x %02x)",
-						pts, (intmax_t)offset, pid, packet[0], packet[1], packet[2], packet[3], packet[4], packet[5], packet[6], packet[7], packet[8], packet[9], packet[10]);
+					eDebug("[eDVBTSTools] getPTS PCR %16llx found at %lld pid %02x (%02x %02x %02x %02x %02x %02x %02x %02x %02x %02x %02x)",
+						pts, (long long)offset, pid, packet[0], packet[1], packet[2], packet[3], packet[4], packet[5], packet[6], packet[7], packet[8], packet[9], packet[10]);
 					if (fixed && fixupPTS(offset, pts))
 						return -1;
 					return 0;
@@ -307,7 +307,7 @@ int eDVBTSTools::getPTS(off_t &offset, pts_t &pts, int fixed)
 			pts |= ((unsigned long long)(payload[13]&0xFE)) >> 1;
 			offset -= 188;
 
-			eDebug("[eDVBTSTools] getPTS PTS %16llx found at %ld pid %02x stream: %02x", pts, (intmax_t)offset, pid, payload[3]);
+			eDebug("[eDVBTSTools] getPTS PTS %16llx found at %lld pid %02x stream: %02x", pts, (long long)offset, pid, payload[3]);
 
 				/* convert to zero-based */
 			if (fixed && fixupPTS(offset, pts))
@@ -463,7 +463,7 @@ int eDVBTSTools::getOffset(off_t &offset, pts_t &pts, int marg)
 			if (p != -1)
 			{
 				pts = p;
-				eDebug("[eDVBTSTools] getOffset aborting. Taking %ld as offset for %lld", (intmax_t)offset, pts);
+				eDebug("[eDVBTSTools] getOffset aborting. Taking %lld as offset for %lld", (long long)offset, pts);
 				return 0;
 			}
 		}
@@ -495,8 +495,8 @@ void eDVBTSTools::calcBegin()
 			pts_t pts = m_pts_begin;
 			if (m_streaminfo.fixupPTS(begin, pts) == 0)
 			{
-				eDebug("[eDVBTSTools] calcBegin [@ML] m_streaminfo.getLastFrame returned %ld, %ld (%us), fixup to: %ld, %lld (%us)",
-				       (intmax_t)m_offset_begin, (intmax_t)m_pts_begin, (unsigned int)(m_pts_begin/90000), (intmax_t)begin, pts, (unsigned int)(pts/90000));
+				eDebug("[eDVBTSTools] calcBegin [@ML] m_streaminfo.getLastFrame returned %lld, %lld (%us), fixup to: %lld, %lld (%us)",
+				       (long long)m_offset_begin, (long long)m_pts_begin, (unsigned int)(m_pts_begin/90000), (long long)begin, pts, (unsigned int)(pts/90000));
 			}
 			m_begin_valid = 1;
 		}
@@ -652,8 +652,8 @@ void eDVBTSTools::takeSamples()
 
 	bytes_per_sample -= bytes_per_sample % m_packet_size;
 
-	eDebug("[eDVBTSTools] takeSamples step %ld, pts begin %lld, pts end %lld, offs begin %ld, offs end %ld:",
-		(intmax_t)bytes_per_sample, m_pts_begin, m_pts_end, (intmax_t)m_offset_begin, (intmax_t)m_offset_end);
+	eDebug("[eDVBTSTools] takeSamples step %lld, pts begin %lld, pts end %lld, offs begin %lld, offs end %lld:",
+		(long long)bytes_per_sample, m_pts_begin, m_pts_end, (long long)m_offset_begin, (long long)m_offset_end);
 
 	for (off_t offset = m_offset_begin; offset < m_offset_end;)
 	{
@@ -696,7 +696,7 @@ int eDVBTSTools::takeSample(off_t off, pts_t &p)
 			}
 		}
 
-		eDebug("[eDVBTSTools] takeSample adding sample %ld: pts %lld -> pos %ld (diff %ld bytes)", (intmax_t)offset_org, p, (intmax_t)off, (intmax_t)(off-offset_org));
+		eDebug("[eDVBTSTools] takeSample adding sample %lld: pts %lld -> pos %lld (diff %lld bytes)", (long long)offset_org, p, (long long)off, (long long)(off-offset_org));
 		m_samples[p] = off;
 		return 0;
 	}
